created and basic designed movie detail component
