+++ src/app/app.module.ts
@@ -14,21 +14,14 @@ import { MenuComponent } from './components/menu/menu.component';
 import { FooterComponent } from './components/footer/footer.component';
 import { CategoryComponent } from './components/category/category.component';
 import { MovieComponent } from './components/movie/movie.component';
-import { MovieListComponent } from './components/movie-list/movie-list.component';
-
-
 import { MiniSliderComponent } from './components/mini-slider/mini-slider.component';
 import { LoginComponent } from './components/login/login.component';
 import { AccountService } from './services/account.service';
 import { LoginGuard } from './login.guard';
-
-
-
 import { AboutComponent } from './components/about/about.component';
 import { AccordionComponent } from './components/accordion/accordion.component';
-;
 import { BigSliderComponent } from './components/big-slider/big-slider.component';
-import { MovieDetailComponent } from './movie-detail/movie-detail.component';
+import { MovieDetailComponent } from './components//movie-detail/movie-detail.component';
 
 
 @NgModule({
@@ -40,7 +33,6 @@ import { MovieDetailComponent } from './movie-detail/movie-detail.component';
     FooterComponent,
     CategoryComponent,
     MovieComponent,
-    MovieListComponent,
     MiniSliderComponent,
     LoginComponent,
     AboutComponent,
