+++ src/app/components/footer/footer.component.html
@@ -4,9 +4,9 @@
         <div class="card-body d-flex flex-column ">
             <div class="justify-content-center d-flex name">{{ dev.name }}</div>
             <ul class="d-flex list-group flex-row justify-content-center">
-                <li><a href="{{ 'mailto:'+dev.mail }}"><i class="far fa-envelope"></i></a></li>
-                <li><a href="{{ dev.github }}"><i class="fab fa-github"></i></a> </li>
-                <li><a href="{{ dev.linkedin }}"><i class="fab fa-linkedin"></i></a></li>
+                <li><a href="{{ 'mailto:'+dev.mail }}" target="_blank"><i class="far fa-envelope"></i></a></li>
+                <li><a href="{{ dev.github }}" target="_blank"><i class="fab fa-github"></i></a> </li>
+                <li><a href="{{ dev.linkedin }}" target="_blank"><i class="fab fa-linkedin"></i></a></li>
             </ul>
         </div>
     </div>
