+++ src/app/components/mini-slider/mini-slider.component.html
@@ -6,7 +6,7 @@
     </div>
     <swiper [config]="config" style="max-height: 150px;">
         <ng-template swiperSlide *ngFor="let movie of popularList">
-            <div class="movie-box">
+            <div class="movie-box" routerLink="movieDetail/{{movie.id}}">
                 <img class="img-fluid" [src]="[imgUrl+movie.backdrop_path]">
                 <div class="about-box">
                     <span> {{movie.title}} </span>
