+++ src/app/components/movie-detail/movie-detail.component.scss
@@ -1,0 +1,18 @@
+@media(min-width:768px) {
+    .movie-container {
+        background: linear-gradient(90deg, #141414 46%, rgba(0, 0, 0, 0) 100%);
+        z-index: 2323;
+    }
+}
+
+.movie-container {
+    .image-container {
+        z-index: -1;
+    }
+    .movie-box {
+        color: #fff;
+        .title {
+            font-size: 22px;
+        }
+    }
+}
