--- src/app/components/movie-list/movie-list.component.html
@@ -1,1 +1,0 @@
-<p>movie-list works!</p>
