--- src/app/movie-detail/movie-detail.component.html
@@ -1,1 +1,0 @@
-<p>movie-detail works!</p>
